Fit the constrained kernel density estimation model with moment constraints

Fit the constrained kernel density estimation model with moment constraints

Usage

fit_constrained_kde(x, h)

Arguments

x

The data points

h

The bandwidth for the kernel density estimation

Value

A list containing the weights, lambda, and other parameters

Examples

set.seed(42)
n <- 100
x <- c(rnorm(0.5 * n, 0, 1), rnorm(0.5 * n, 1.5, 1/3))
h <- 0.6
model <- fit_constrained_kde(x, h)
print(model)
#> $weights
#>   [1]  0.0078605175  0.0080569977  0.0155133830  0.0117185057  0.0148842648
#>   [6]  0.0173101055  0.0081636641  0.0174791588  0.0255743500  0.0178931155
#>  [11]  0.0078507286 -0.0164717540  0.0017944913  0.0138846724  0.0168698978
#>  [16]  0.0116820769  0.0137618122  0.0004006985  0.0004936255  0.0078464275
#>  [21]  0.0132571567  0.0010388140  0.0161628335  0.0079513689  0.0141420378
#>  [26]  0.0105529403  0.0143656639  0.0010632902  0.0140422106  0.0069066938
#>  [31]  0.0141112366  0.0109198753  0.0085152124  0.0073597592  0.0133901931
#>  [36]  0.0011290818  0.0051622691  0.0045331783  0.0005068208  0.0185757093
#>  [41]  0.0176899961  0.0120388655  0.0103957554  0.0057926141  0.0018517212
#>  [46]  0.0144505535  0.0048956198  0.0079647104  0.0105328292  0.0114542567
#>  [51]  0.0086681420  0.0079122948  0.0269488636  0.0096973458  0.0082381263
#>  [56]  0.0085674597  0.0098594541  0.0082382300  0.0134279376  0.0085852807
#>  [61]  0.0078657168  0.0083904784  0.0094498399  0.0187236455  0.0078878067
#>  [66]  0.0162343854  0.0087008712  0.0123367036  0.0113015025  0.0100595913
#>  [71]  0.0080929758  0.0080281386  0.0096156050  0.0080179806  0.0078466021
#>  [76]  0.0094466843  0.0103084919  0.0090464786  0.0079700472  0.0081472051
#>  [81]  0.0231660441  0.0085286650  0.0082362333  0.0080012064  0.0082494438
#>  [86]  0.0095684541  0.0079318761  0.0079540983  0.0113993722  0.0106213011
#>  [91]  0.0184941843  0.0078468709  0.0097295952  0.0184644042  0.0081583519
#>  [96]  0.0079542879  0.0081801159  0.0086152789  0.0082242408  0.0097420837
#> 
#> $lambda
#>             [,1]
#> [1,]  0.37704674
#> [2,] -0.04480643
#> [3,]  1.20835969
#> [4,] -0.58058765
#> 
#> $x
#>   [1]  1.37095845 -0.56469817  0.36312841  0.63286260  0.40426832 -0.10612452
#>   [7]  1.51152200 -0.09465904  2.01842371 -0.06271410  1.30486965  2.28664539
#>  [13] -1.38886070 -0.27878877 -0.13332134  0.63595040 -0.28425292 -2.65645542
#>  [19] -2.44046693  1.32011335 -0.30663859 -1.78130843 -0.17191736  1.21467470
#>  [25]  1.89519346 -0.43046913 -0.25726938 -1.76316309  0.46009735 -0.63999488
#>  [31]  0.45545012  0.70483734  1.03510352 -0.60892638  0.50495512 -1.71700868
#>  [37] -0.78445901 -0.85090759 -2.41420765  0.03612261  0.20599860 -0.36105730
#>  [43]  0.75816324 -0.72670483 -1.36828104  0.43281803 -0.81139318  1.44410126
#>  [49] -0.43144620  0.65564788  1.60730842  1.23872035  2.02524251  1.71429977
#>  [55]  1.52992022  1.59218358  1.72642961  1.52994430  0.50230331  1.59496098
#>  [61]  1.37758845  1.56174352  1.69394124  1.96657894  1.25756931  1.93418088
#>  [67]  1.61194937  1.84616870  1.80690952  1.74029272  1.15229369  1.46993787
#>  [73]  1.70783939  1.18215888  1.31905706  1.69366550  1.75605958  1.65458920
#>  [79]  1.20474123  1.13340637  2.00423567  1.58597381  1.52948008  1.45970115
#>  [85]  1.10189037  1.70399897  1.42762005  1.43908110  1.81111544  1.77392437
#>  [91]  1.96403879  1.34127536  1.71678285  1.96370349  1.12973704  1.21306914
#>  [97]  1.12275377  1.01359533  1.52666085  1.71773478
#> 
#> $h
#> [1] 0.6
#>