Predict density and moments using the fitted constrained KDE model

Predict density and moments using the fitted constrained KDE model

Usage

predict_constrained_kde(model, x_seq)

Arguments

model

The fitted model object

x_seq

The sequence of points to estimate the density

Value

A list containing density estimates and first four moments

Examples

set.seed(42)
n <- 100
x <- c(rnorm(0.5*n, 0, 1), rnorm(0.5 * n, 1.5, 1/3))
h <- 0.6
model <- fit_constrained_kde(x, h)
x_seq <- seq(min(x) - 1, max(x) + 1, length.out = 1000)
predictions <- predict_constrained_kde(model, x_seq)
print(predictions$density)
#>    [1]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>    [6]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [11]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [16]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [21]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [26]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [31]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [36]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [41]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [46]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [51]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [56]  0.000000e+00  0.000000e+00  0.000000e+00  6.663401e-08  6.913206e-07
#>   [61]  1.954582e-06  3.833590e-06  6.305788e-06  9.348891e-06  1.294088e-05
#>   [66]  1.706001e-05  2.168482e-05  2.679408e-05  3.236688e-05  3.838254e-05
#>   [71]  4.482068e-05  5.166117e-05  5.888416e-05  6.647007e-05  7.439959e-05
#>   [76]  8.265367e-05  9.121356e-05  1.000607e-04  1.091770e-04  1.185444e-04
#>   [81]  1.281452e-04  1.379620e-04  1.479777e-04  1.581753e-04  1.685383e-04
#>   [86]  1.790504e-04  1.896955e-04  2.004578e-04  2.113217e-04  2.223282e-04
#>   [91]  2.340576e-04  2.466325e-04  2.600104e-04  2.742967e-04  2.900639e-04
#>   [96]  3.073134e-04  3.259762e-04  3.459844e-04  3.672707e-04  3.897691e-04
#>  [101]  4.134143e-04  4.381420e-04  4.638890e-04  4.905928e-04  5.181921e-04
#>  [106]  5.466263e-04  5.758359e-04  6.057623e-04  6.363479e-04  6.675360e-04
#>  [111]  6.992708e-04  7.314975e-04  7.641622e-04  7.972121e-04  8.305951e-04
#>  [116]  8.642603e-04  8.981575e-04  9.322377e-04  9.664526e-04  1.000755e-03
#>  [121]  1.035099e-03  1.069438e-03  1.103729e-03  1.137928e-03  1.171992e-03
#>  [126]  1.205881e-03  1.239552e-03  1.272968e-03  1.306088e-03  1.338875e-03
#>  [131]  1.371293e-03  1.403305e-03  1.434876e-03  1.465973e-03  1.496562e-03
#>  [136]  1.526612e-03  1.556091e-03  1.584968e-03  1.613216e-03  1.640804e-03
#>  [141]  1.667707e-03  1.693897e-03  1.719350e-03  1.744039e-03  1.767943e-03
#>  [146]  1.791038e-03  1.813302e-03  1.834716e-03  1.855258e-03  1.874911e-03
#>  [151]  1.893656e-03  1.911476e-03  1.928355e-03  1.944278e-03  1.959232e-03
#>  [156]  1.973202e-03  1.986177e-03  1.998145e-03  2.009096e-03  2.019021e-03
#>  [161]  2.027911e-03  2.035758e-03  2.042557e-03  2.048301e-03  2.052985e-03
#>  [166]  2.056607e-03  2.059163e-03  2.060651e-03  2.061070e-03  2.060422e-03
#>  [171]  2.058705e-03  2.055924e-03  2.052079e-03  2.047176e-03  2.041219e-03
#>  [176]  2.034214e-03  2.026167e-03  2.017086e-03  2.006980e-03  1.995858e-03
#>  [181]  1.983731e-03  1.970609e-03  1.956506e-03  1.941435e-03  1.925650e-03
#>  [186]  1.910441e-03  1.895959e-03  1.882904e-03  1.872215e-03  1.863799e-03
#>  [191]  1.857558e-03  1.853398e-03  1.851226e-03  1.851023e-03  1.854003e-03
#>  [196]  1.860519e-03  1.870421e-03  1.883566e-03  1.899812e-03  1.919021e-03
#>  [201]  1.941059e-03  1.965792e-03  1.993092e-03  2.022834e-03  2.054895e-03
#>  [206]  2.089155e-03  2.125498e-03  2.163811e-03  2.203984e-03  2.245910e-03
#>  [211]  2.289485e-03  2.334608e-03  2.381182e-03  2.429113e-03  2.478309e-03
#>  [216]  2.528681e-03  2.580145e-03  2.632620e-03  2.686025e-03  2.740286e-03
#>  [221]  2.795330e-03  2.851087e-03  2.907492e-03  2.964480e-03  3.021993e-03
#>  [226]  3.079973e-03  3.138366e-03  3.197122e-03  3.256193e-03  3.315535e-03
#>  [231]  3.375106e-03  3.434659e-03  3.493694e-03  3.552138e-03  3.609941e-03
#>  [236]  3.667051e-03  3.723421e-03  3.779008e-03  3.833771e-03  3.887670e-03
#>  [241]  3.940675e-03  3.994409e-03  4.050080e-03  4.107575e-03  4.168575e-03
#>  [246]  4.234123e-03  4.303998e-03  4.377979e-03  4.455856e-03  4.537420e-03
#>  [251]  4.622469e-03  4.710806e-03  4.802241e-03  4.896585e-03  4.993660e-03
#>  [256]  5.093288e-03  5.195301e-03  5.299532e-03  5.405823e-03  5.514019e-03
#>  [261]  5.623970e-03  5.735534e-03  5.848363e-03  5.961719e-03  6.075411e-03
#>  [266]  6.189287e-03  6.302828e-03  6.415388e-03  6.526798e-03  6.636897e-03
#>  [271]  6.745526e-03  6.852532e-03  6.957767e-03  7.061087e-03  7.162354e-03
#>  [276]  7.261432e-03  7.358192e-03  7.452508e-03  7.544260e-03  7.633333e-03
#>  [281]  7.719614e-03  7.802998e-03  7.883382e-03  7.960669e-03  8.034767e-03
#>  [286]  8.105587e-03  8.173045e-03  8.237064e-03  8.297569e-03  8.354491e-03
#>  [291]  8.407764e-03  8.457329e-03  8.503130e-03  8.545117e-03  8.583242e-03
#>  [296]  8.617464e-03  8.647747e-03  8.674059e-03  8.696370e-03  8.714659e-03
#>  [301]  8.728907e-03  8.739100e-03  8.745229e-03  8.747290e-03  8.745282e-03
#>  [306]  8.739212e-03  8.729087e-03  8.714923e-03  8.696738e-03  8.674555e-03
#>  [311]  8.648403e-03  8.618313e-03  8.584325e-03  8.546478e-03  8.504821e-03
#>  [316]  8.459404e-03  8.410284e-03  8.357520e-03  8.302808e-03  8.251570e-03
#>  [321]  8.204067e-03  8.160123e-03  8.119564e-03  8.082229e-03  8.051797e-03
#>  [326]  8.032216e-03  8.023072e-03  8.023993e-03  8.039600e-03  8.072778e-03
#>  [331]  8.122855e-03  8.189174e-03  8.271092e-03  8.367981e-03  8.479228e-03
#>  [336]  8.604232e-03  8.745387e-03  8.908375e-03  9.092528e-03  9.296978e-03
#>  [341]  9.520872e-03  9.763377e-03  1.002368e-02  1.030098e-02  1.059451e-02
#>  [346]  1.090349e-02  1.122720e-02  1.156491e-02  1.191647e-02  1.228932e-02
#>  [351]  1.268517e-02  1.310299e-02  1.354174e-02  1.400474e-02  1.449855e-02
#>  [356]  1.502193e-02  1.557355e-02  1.615142e-02  1.675314e-02  1.737885e-02
#>  [361]  1.803588e-02  1.872383e-02  1.944092e-02  2.018544e-02  2.095568e-02
#>  [366]  2.174997e-02  2.256632e-02  2.340169e-02  2.425432e-02  2.512258e-02
#>  [371]  2.600490e-02  2.689975e-02  2.780561e-02  2.872101e-02  2.964448e-02
#>  [376]  3.057462e-02  3.151003e-02  3.244937e-02  3.339235e-02  3.436114e-02
#>  [381]  3.536383e-02  3.639800e-02  3.746124e-02  3.855123e-02  3.966565e-02
#>  [386]  4.080226e-02  4.195886e-02  4.313327e-02  4.432371e-02  4.554092e-02
#>  [391]  4.678911e-02  4.806562e-02  4.936783e-02  5.069318e-02  5.203918e-02
#>  [396]  5.340335e-02  5.478463e-02  5.619653e-02  5.764078e-02  5.911456e-02
#>  [401]  6.063016e-02  6.220938e-02  6.385297e-02  6.555731e-02  6.733439e-02
#>  [406]  6.918619e-02  7.110784e-02  7.309452e-02  7.514154e-02  7.724427e-02
#>  [411]  7.939819e-02  8.159887e-02  8.384196e-02  8.612296e-02  8.843545e-02
#>  [416]  9.077470e-02  9.314873e-02  9.556496e-02  9.801799e-02  1.005030e-01
#>  [421]  1.030152e-01  1.055522e-01  1.081303e-01  1.107488e-01  1.134025e-01
#>  [426]  1.160897e-01  1.188267e-01  1.216214e-01  1.244860e-01  1.274138e-01
#>  [431]  1.303984e-01  1.334342e-01  1.365352e-01  1.397027e-01  1.429296e-01
#>  [436]  1.462090e-01  1.495340e-01  1.528981e-01  1.562948e-01  1.597176e-01
#>  [441]  1.631604e-01  1.666172e-01  1.700820e-01  1.735490e-01  1.770127e-01
#>  [446]  1.804676e-01  1.839246e-01  1.873924e-01  1.908647e-01  1.943358e-01
#>  [451]  1.977998e-01  2.012510e-01  2.046842e-01  2.080938e-01  2.114749e-01
#>  [456]  2.148224e-01  2.181314e-01  2.213974e-01  2.246159e-01  2.277824e-01
#>  [461]  2.308928e-01  2.339430e-01  2.369293e-01  2.398478e-01  2.426950e-01
#>  [466]  2.454676e-01  2.481623e-01  2.507760e-01  2.533058e-01  2.557490e-01
#>  [471]  2.581080e-01  2.604018e-01  2.626296e-01  2.647885e-01  2.668753e-01
#>  [476]  2.688873e-01  2.708220e-01  2.726767e-01  2.744494e-01  2.761378e-01
#>  [481]  2.777401e-01  2.792544e-01  2.806793e-01  2.820132e-01  2.832549e-01
#>  [486]  2.844034e-01  2.854577e-01  2.864171e-01  2.872811e-01  2.880492e-01
#>  [491]  2.887212e-01  2.892932e-01  2.897614e-01  2.901380e-01  2.904361e-01
#>  [496]  2.906550e-01  2.907942e-01  2.908474e-01  2.908125e-01  2.907029e-01
#>  [501]  2.905284e-01  2.902817e-01  2.899606e-01  2.895755e-01  2.891386e-01
#>  [506]  2.886488e-01  2.881111e-01  2.875533e-01  2.869858e-01  2.864006e-01
#>  [511]  2.857921e-01  2.851585e-01  2.844983e-01  2.838238e-01  2.831618e-01
#>  [516]  2.825120e-01  2.818720e-01  2.812398e-01  2.806136e-01  2.799918e-01
#>  [521]  2.793730e-01  2.787536e-01  2.781241e-01  2.774825e-01  2.768281e-01
#>  [526]  2.761599e-01  2.754694e-01  2.747518e-01  2.740068e-01  2.732339e-01
#>  [531]  2.724333e-01  2.716054e-01  2.707636e-01  2.699187e-01  2.690705e-01
#>  [536]  2.682256e-01  2.673944e-01  2.665763e-01  2.657704e-01  2.649763e-01
#>  [541]  2.641937e-01  2.634226e-01  2.626691e-01  2.619453e-01  2.612515e-01
#>  [546]  2.605879e-01  2.599552e-01  2.593540e-01  2.587855e-01  2.582509e-01
#>  [551]  2.577626e-01  2.573221e-01  2.569040e-01  2.565055e-01  2.561271e-01
#>  [556]  2.557697e-01  2.554343e-01  2.551223e-01  2.548352e-01  2.545748e-01
#>  [561]  2.543431e-01  2.541397e-01  2.539515e-01  2.537782e-01  2.536220e-01
#>  [566]  2.534853e-01  2.533710e-01  2.532817e-01  2.532207e-01  2.531861e-01
#>  [571]  2.531650e-01  2.531593e-01  2.531699e-01  2.531770e-01  2.531630e-01
#>  [576]  2.531289e-01  2.530732e-01  2.529799e-01  2.528491e-01  2.526829e-01
#>  [581]  2.524836e-01  2.522538e-01  2.519963e-01  2.517139e-01  2.514099e-01
#>  [586]  2.510874e-01  2.507510e-01  2.504149e-01  2.500839e-01  2.497427e-01
#>  [591]  2.493973e-01  2.490538e-01  2.487150e-01  2.483838e-01  2.480551e-01
#>  [596]  2.477121e-01  2.473564e-01  2.469905e-01  2.466070e-01  2.461924e-01
#>  [601]  2.457385e-01  2.452372e-01  2.446927e-01  2.441188e-01  2.435298e-01
#>  [606]  2.429085e-01  2.422527e-01  2.415718e-01  2.408669e-01  2.401372e-01
#>  [611]  2.393822e-01  2.386094e-01  2.378232e-01  2.370231e-01  2.362136e-01
#>  [616]  2.354068e-01  2.346212e-01  2.338581e-01  2.331143e-01  2.323722e-01
#>  [621]  2.316301e-01  2.308865e-01  2.301485e-01  2.294175e-01  2.286915e-01
#>  [626]  2.279686e-01  2.272472e-01  2.265260e-01  2.258051e-01  2.250933e-01
#>  [631]  2.243927e-01  2.237209e-01  2.230818e-01  2.224739e-01  2.219047e-01
#>  [636]  2.213765e-01  2.208877e-01  2.204371e-01  2.200297e-01  2.196782e-01
#>  [641]  2.193874e-01  2.191561e-01  2.189831e-01  2.188548e-01  2.187529e-01
#>  [646]  2.186757e-01  2.186264e-01  2.186125e-01  2.186425e-01  2.187294e-01
#>  [651]  2.188738e-01  2.190828e-01  2.193624e-01  2.197203e-01  2.201551e-01
#>  [656]  2.206652e-01  2.212494e-01  2.219065e-01  2.226394e-01  2.234567e-01
#>  [661]  2.243604e-01  2.253844e-01  2.265444e-01  2.278382e-01  2.292640e-01
#>  [666]  2.308211e-01  2.324994e-01  2.342923e-01  2.361976e-01  2.382211e-01
#>  [671]  2.403802e-01  2.426856e-01  2.451238e-01  2.477047e-01  2.504280e-01
#>  [676]  2.532893e-01  2.562678e-01  2.593535e-01  2.625430e-01  2.658281e-01
#>  [681]  2.691813e-01  2.725929e-01  2.760581e-01  2.795725e-01  2.831361e-01
#>  [686]  2.867670e-01  2.904493e-01  2.941799e-01  2.979855e-01  3.018771e-01
#>  [691]  3.058568e-01  3.099221e-01  3.140755e-01  3.183115e-01  3.226345e-01
#>  [696]  3.270379e-01  3.315221e-01  3.360880e-01  3.407274e-01  3.454327e-01
#>  [701]  3.501968e-01  3.550292e-01  3.599415e-01  3.649270e-01  3.699757e-01
#>  [706]  3.750543e-01  3.801471e-01  3.852607e-01  3.903753e-01  3.954765e-01
#>  [711]  4.005558e-01  4.056053e-01  4.106172e-01  4.155873e-01  4.205257e-01
#>  [716]  4.254156e-01  4.302425e-01  4.349987e-01  4.396769e-01  4.442815e-01
#>  [721]  4.488207e-01  4.532870e-01  4.576715e-01  4.619685e-01  4.662337e-01
#>  [726]  4.704694e-01  4.746662e-01  4.788148e-01  4.829064e-01  4.869465e-01
#>  [731]  4.909502e-01  4.949228e-01  4.988979e-01  5.028945e-01  5.069021e-01
#>  [736]  5.109104e-01  5.149090e-01  5.188882e-01  5.228386e-01  5.267514e-01
#>  [741]  5.306178e-01  5.344299e-01  5.381798e-01  5.418601e-01  5.454640e-01
#>  [746]  5.489849e-01  5.524167e-01  5.557537e-01  5.589905e-01  5.621222e-01
#>  [751]  5.651443e-01  5.680528e-01  5.708438e-01  5.735141e-01  5.760609e-01
#>  [756]  5.784816e-01  5.807741e-01  5.829367e-01  5.849682e-01  5.868640e-01
#>  [761]  5.886129e-01  5.902141e-01  5.916643e-01  5.929532e-01  5.940798e-01
#>  [766]  5.950446e-01  5.958486e-01  5.964930e-01  5.969796e-01  5.973098e-01
#>  [771]  5.974678e-01  5.974490e-01  5.972503e-01  5.968691e-01  5.963099e-01
#>  [776]  5.955705e-01  5.946403e-01  5.935069e-01  5.921730e-01  5.906395e-01
#>  [781]  5.889020e-01  5.869655e-01  5.848353e-01  5.825157e-01  5.800028e-01
#>  [786]  5.773006e-01  5.744153e-01  5.713445e-01  5.680796e-01  5.646116e-01
#>  [791]  5.609466e-01  5.570907e-01  5.530411e-01  5.488018e-01  5.443776e-01
#>  [796]  5.397659e-01  5.349734e-01  5.300077e-01  5.248770e-01  5.195897e-01
#>  [801]  5.141545e-01  5.085760e-01  5.028548e-01  4.969933e-01  4.909819e-01
#>  [806]  4.848282e-01  4.785394e-01  4.721149e-01  4.655630e-01  4.588933e-01
#>  [811]  4.521152e-01  4.452287e-01  4.382310e-01  4.311292e-01  4.239337e-01
#>  [816]  4.166549e-01  4.093038e-01  4.018915e-01  3.944292e-01  3.869201e-01
#>  [821]  3.793695e-01  3.717746e-01  3.641363e-01  3.564637e-01  3.487582e-01
#>  [826]  3.410223e-01  3.332614e-01  3.254873e-01  3.177116e-01  3.099465e-01
#>  [831]  3.022044e-01  2.944898e-01  2.868096e-01  2.791664e-01  2.715330e-01
#>  [836]  2.639144e-01  2.563222e-01  2.487680e-01  2.412586e-01  2.337971e-01
#>  [841]  2.263952e-01  2.190641e-01  2.118039e-01  2.046046e-01  1.974687e-01
#>  [846]  1.903943e-01  1.833783e-01  1.764305e-01  1.695622e-01  1.627845e-01
#>  [851]  1.561092e-01  1.495460e-01  1.430952e-01  1.367667e-01  1.305720e-01
#>  [856]  1.245234e-01  1.186329e-01  1.129014e-01  1.073194e-01  1.018807e-01
#>  [861]  9.657229e-02  9.136970e-02  8.626710e-02  8.126740e-02  7.636872e-02
#>  [866]  7.157384e-02  6.688448e-02  6.229880e-02  5.782414e-02  5.345599e-02
#>  [871]  4.919836e-02  4.505940e-02  4.104738e-02  3.716750e-02  3.340462e-02
#>  [876]  2.975447e-02  2.622446e-02  2.282209e-02  1.955501e-02  1.642186e-02
#>  [881]  1.341824e-02  1.055131e-02  7.828380e-03  5.256859e-03  2.844252e-03
#>  [886]  5.981648e-04 -1.482918e-03 -3.406247e-03 -5.164938e-03 -6.751842e-03
#>  [891] -8.159721e-03 -9.382224e-03 -1.043120e-02 -1.130734e-02 -1.200402e-02
#>  [896] -1.251454e-02 -1.286899e-02 -1.311741e-02 -1.325860e-02 -1.328887e-02
#>  [901] -1.320450e-02 -1.300242e-02 -1.270456e-02 -1.232159e-02 -1.187880e-02
#>  [906] -1.141989e-02 -1.096101e-02 -1.050270e-02 -1.004556e-02 -9.590146e-03
#>  [911] -9.137068e-03 -8.686926e-03 -8.240336e-03 -7.797926e-03 -7.360333e-03
#>  [916] -6.928206e-03 -6.502207e-03 -6.083005e-03 -5.671285e-03 -5.267740e-03
#>  [921] -4.873074e-03 -4.488004e-03 -4.113257e-03 -3.749570e-03 -3.397694e-03
#>  [926] -3.058388e-03 -2.732425e-03 -2.420586e-03 -2.123666e-03 -1.842470e-03
#>  [931] -1.577814e-03 -1.330525e-03 -1.101440e-03 -8.914107e-04 -7.012962e-04
#>  [936] -5.319685e-04 -3.843104e-04 -2.592158e-04 -1.575897e-04 -8.034815e-05
#>  [941] -2.841853e-05 -2.739164e-06  0.000000e+00  0.000000e+00  0.000000e+00
#>  [946]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [951]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [956]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [961]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [966]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [971]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [976]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [981]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [986]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [991]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [996]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
print(predictions$moments)
#> [1] 0.939022 1.700662 2.721888 5.162648